requirements.txt
discord
flask
openai
python-dotenv